Refactor next boilerplate and add gatsby to cli
